requirements.txt
pymongo==4.10.1
schedule==1.1.0
python-dotenv
